Card background letters -> SVG
Though it feels kinda heavy, the previous method relied on a font that we
weren't shipping yet. So this is actually quite a bit lighter.

I'm not happy that the *fill* and *opacity* are declared inside the SVG, but
from what I can tell, CSS cannot be applied to SVG unless the SVG is in the HTML
document, and I wanted to use the SVGs as a background... etc.

Huge, and pleasant surprise that multiple background images can be applied to
the same element, overlayed, and independently positioned. Made my day. :-)

Letters are created using Roboto Mono (font size 632.8 for 600px tall, 316.4 for
300px).

SVGs were generated by Inkscape and cleaned-up by hand.

Fix #6