add giraffe image to "resources" card
I initially thought that this image was provided under the CC0 license. The
creator, Ryan McGuire, used to use the CC0 license, but as of 2017, he's started
using a custom license. The giraffe image was created after the license change
so it falls under the new custom license (see the copyright page for a link to
the custom license terms).