Replace Giraffe image with Floppies
First whack at it.

The motivation is licensing issues. While we /can/ use the Giraffe image, the
license is very problematic. The author has not responded to my email requesting
clarification, so I am choosing to remove the image. Despite what his license
intends, it in-effect doesn't allow any usage. And at best, does not allow
others to re-use or build-upon our version of the image. That's pretty much the
anti of what we do, so I'm just replacing it outright.