Move SVGs from CSS file to HTML
Bloating the CSS file like that was needlessly slowing down the time to the
first paint. Putting it in the HTML gives the browser the context it needs to
make faster decisions.

Also saves loading them on each and every page.

Not thrilled about the amount of positioning I need to do here (for the sake of
vertical stacking), but it's worth the trade-off.