hack around lack of "break-after: avoid" support; make index validate again
Since there's no "break-after/before avoid support, it is easiest to just wrap
the <dt> and <dd> together in a <div>. This /is/ legal in HTML5, as <dl> does
accept <div> as direct children.

As for validation: I had forgotten to close many of the <dd> tags on the index
page, and yet neither of the W3C's HTML5 validators caught it. Odd.